Move Encryption class into common module
We'll have to beable to access the class from the web-app module to
encrypt the ssh secret-key used to access the user FileStore.
