+++ datavault-worker/src/main/java/org/datavaultplatform/worker/tasks/Retrieve.java
@@ -9,9 +9,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import javax.crypto.Cipher;
-import javax.crypto.SecretKey;
-
 import org.apache.commons.io.FileUtils;
 import org.datavaultplatform.common.event.Error;
 import org.datavaultplatform.common.event.InitStates;
@@ -25,11 +22,10 @@ import org.datavaultplatform.common.storage.UserStore;
 import org.datavaultplatform.common.storage.Verify;
 import org.datavaultplatform.common.task.Context;
 import org.datavaultplatform.common.task.Task;
-import org.datavaultplatform.common.task.Context.AESMode;
 import org.datavaultplatform.worker.operations.Packager;
 import org.datavaultplatform.worker.operations.ProgressTracker;
 import org.datavaultplatform.worker.operations.Tar;
-import org.datavaultplatform.worker.operations.Encryption;
+import org.datavaultplatform.common.crypto.Encryption;
 import org.datavaultplatform.worker.operations.FileSplitter;
 import org.datavaultplatform.worker.queue.EventSender;
 import org.slf4j.Logger;
